feat: enhance ban status handling with caching and improve report user functionality
