+++ src/handleReportUser.ts
@@ -1,6 +1,6 @@
 import { Context, MenuItemOnPressEvent, JSONObject, FormOnSubmitEvent, FormFunction, TriggerContext } from "@devvit/public-api";
 import { CONTROL_SUBREDDIT } from "./constants.js";
-import { getPostOrCommentById, getUserOrUndefined, isModeratorWithCache } from "./utility.js";
+import { getPostOrCommentById, getUserOrUndefined, isBannedWithCache, isModeratorWithCache } from "./utility.js";
 import { getUserStatus, UserStatus } from "./dataStore.js";
 import { addExternalSubmissionFromClientSub } from "./externalSubmissions.js";
 import { queryForm, reportForm } from "./main.js";
@@ -11,7 +11,6 @@ import { recordReportForSummary } from "./modmail/actionSummary.js";
 import { canUserReceiveFeedback } from "./submissionFeedback.js";
 import { isLinkId } from "@devvit/public-api/types/tid.js";
 import { addClassificationQueryToQueue } from "./modmail/classificationQuery.js";
-import { isBanned } from "devvit-helpers";
 
 enum ReportFormField {
     ReportContext = "reportContext",
@@ -20,6 +19,8 @@ enum ReportFormField {
 }
 
 export const reportFormDefinition: FormFunction = data => ({
+    title: `Report u/${data.username} to Bot Bouncer`,
+    description: "Thank you for reporting this user. Our team will review the account manually and take appropriate action.",
     fields: [
         {
             type: "paragraph",
@@ -125,7 +126,7 @@ export async function handleReportUser (event: MenuItemOnPressEvent, context: Co
         return;
     }
 
-    if (await isBanned(context.reddit, currentUser.username, CONTROL_SUBREDDIT)) {
+    if (await isBannedWithCache(currentUser.username, context, CONTROL_SUBREDDIT)) {
         context.ui.showToast("You are currently banned from r/BotBouncer, so you cannot report other users.");
         return;
     }
@@ -139,6 +140,7 @@ export async function handleReportUser (event: MenuItemOnPressEvent, context: Co
 
     const canReceiveFeedback = await canUserReceiveFeedback(currentUser.username, context);
     const data = {
+        username: target.authorName,
         feedbackHelpText: canReceiveFeedback ? "You must be able to receive chat messages from /u/bot-bouncer to receive this notification" : "We've tried to send feedback for you several times but this hasn't worked. Check to make sure you can receive chats from /u/bot-bouncer. This option will return within 24h.",
         feedbackDisabled: !canReceiveFeedback,
     };
